fix system detail property view
