--- src/modules/systems/components/Description.tsx
@@ -1,6 +1,0 @@
-const Description = ({ data }) => {
-  const { description } = data
-  return <p className="px-3 w-full">{description}</p>
-}
-
-export default Description
