--- src/modules/systems/components/Title.tsx
@@ -1,3 +1,0 @@
-const Title = ({ data }) => <h1 className="text-2xl">{data.name}</h1>
-
-export default Title
