+++ src/modules/systems/components/relationsSection/RelationsSection.tsx
@@ -10,9 +10,9 @@ import { useEndpoint } from '@/hooks/useEndpoint'
 import { message } from '@/i18n/src/messages'
 import { RELATION_TYPE_CODE } from '@/modules/systems/types/constants'
 import { ModalButtons } from '@/types/form'
-import { SystemRelationshipResponse } from '@/types/responses'
 
 import { useRelationMapRows } from '../../hooks/relations/useMapRows'
+import { SystemRelationshipResponse } from '../../types/responses'
 import AddRelationForm from './AddRelationForm'
 
 const messages = message.systemsPage.relations
